Add stage_blockhashes header
